+++ app/src/component/bookmarks/bookmarks.vue
@@ -345,6 +345,9 @@ module.exports = {
     databaseManager.updateListing();
 
     $(document).ready(function() {
+      // Hacky workaround to fix the loading animation inside the textbox
+      $(".ivu-icon-load-c").hide();
+
       $(".ivu-input").keyup(function(event) {
         self.filterList();
       });
@@ -378,7 +381,8 @@ module.exports = {
         self.wordList = self.originalWordList;
       } else {
         self.wordList = self.originalWordList.filter((word) => {
-          if(word.definitions.join().includes(self.filterTerm) || word.pinyin.includes(self.filterTerm) || word.simplified.includes(self.filterTerm) || word.traditional.includes(self.filterTerm)) {
+          const term = self.filterTerm.toLowerCase();
+          if(word.definitions.join().toLowerCase().includes(term) || word.pinyin.toLowerCase().includes(term) || word.simplified.includes(term) || word.traditional.includes(term)) {
             return word;
           }
         });