+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.0] - 2018-08-12
+### Fixed
+- Fixed loading animation in all text fields as described in Issue #103
+
+### Added
+- Added help menu as described in Issue #72
+- Added clipboard and web reader as described in Issue #60
+
+### Changed
+- Updated to a newer version of CC-CEDICT (Released 2018-08-12)
+
 ## [1.2.0] - 2018-05-12
 ### Added
 - Added very basic pinyin detection for search as described in #26